refactor(Phonemes): :truck: Rename infer_phonemes endpoint
