+++ app/schemas/audio_phonemes.py
@@ -6,6 +6,5 @@ from pydantic import BaseModel
 # No request for now, it's just a wav file
 # Could potentially include language in the future
 
-class PhoneResponse(BaseModel):
-    phones: List[str]
-    confidence: List[float]
+class InferPhonemesResponse(BaseModel):
+    phonemes: List[str]
